server_fastapi/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
motor==3.6.0
beanie==1.27.0
pymongo==4.9.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Google Generative AI
google-generativeai==0.8.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
pypdf==5.1.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# HTTP Client
httpx==0.27.2
requests==2.32.3

# Scheduling
APScheduler==3.10.4

# CORS
fastapi-cors==0.0.6
